requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
keyrings.cryptfile>=1.3.9
# 台新 SDK（.whl 手動安裝）：
#   下載：https://ml-fugle-api.tssco.com.tw/FugleSDK/docs/trading/library/python/download
#   pip install taishin_sdk-<version>-<platform>.whl
# 玉山 SDK（.whl 手動安裝）：
#   下載：https://www.esunsec.com.tw/trading-platforms/api-trading/docs/download/download-sdk/
#   pip install esun_trade-<version>-<platform>.whl
